Refactoriza lógica y mejora validaciones en endpoints
Se realizaron los siguientes cambios principales:

- Eliminación de código muerto (`EjecutarConsultaAsync`).
- Consolidación de lógica duplicada en helpers reutilizables:
  `InsertarUnDetFleteAsync` e `InsertarMarcaFleteAsync`.
- Validaciones adicionales:
  - Coordenadas GPS inválidas (0,0) rechazadas en endpoints.
  - Parámetros obligatorios y rangos validados en varios endpoints.
- Mejoras en transacciones y atomicidad:
  - Operaciones críticas envueltas en transacciones.
  - Rollback completo en caso de fallos en `SincronizarFletes`.
- Optimización de queries y filtros para evitar escaneos masivos.
- Correcciones de lógica de negocio:
  - Estado `'F'` al finalizar fletes.
  - Restricción de reanudación a estados permitidos (`C`, `P`).
- Consistencia en la API:
  - Respuestas envueltas en `ApiResponse`.
  - Campo derivado `EsPendiente` añadido en respuestas.
- Correcciones de tipos (`double` a `decimal` en coordenadas).
- Uso de `GETDATE()` en lugar de `DateTime.Now` para evitar problemas
  de zonas horarias.
- Documentación y comentarios añadidos para facilitar mantenimiento.

Estos cambios mejoran la calidad del código, la consistencia de la API
y la experiencia del cliente, además de optimizar el rendimiento.

+++ BusCheckInV2API/Models/DetFleteRequest.cs
@@ -4,8 +4,8 @@
     {
         public int IdFletePer { get; set; }
         public int FlePer_CveNomina { get; set; }
-        public double FlePer_Latitud { get; set; }
-        public double FlePer_Longitud { get; set; }
+        public decimal FlePer_Latitud { get; set; }
+        public decimal FlePer_Longitud { get; set; }
         public DateTime FlePer_Fecha { get; set; }
         public string? FlePer_Nombre { get; set; }
     }
